Add link to 'My boards' on front page, refactor Navbar
